[upnp] Use the service type from the right location
If we do an announce to discover a device, it may respond
with a 'WANPPP' service type in it's udp response. When
we analyze the services for that device, it might only
advertise 'WANIP' as a service.

As such, we should choose the service type based on what
is actually advertised. This makes the library work with
my virgin media router.