Some cleanups - maybe another release?
I hid some API so only the necessary stuff is public.

I also removed the IMapper stuff and exposed the exact
same functionality in ISearcher and NatUtility. Instead
of mapping, you just search and pass in the IP of the
gateway. It runs through the exact same codepath then!

The main INatDevice interface contains only the Async
methods, and the synchronous ones, and ones using
BeginX/EndX are declared as extension methods instead.

I'm in two minds about just removing them completely,
but they kinda do still serve a bit of a purpose.