[core] Issue SSDP requests for 'all' and for specific URNS
Some routers respond to 'all' request and list their
WANPPPConnection or WANIPConnection services, some do
not.

For the ones which do not, querying for
InternetGatewayDevice and accepting that as a valid
possibility should be sufficient to allow Mono.Nat
to fetch the service list and get confirmation
that WANIP or WANPPP is supported.

Fixes https://github.com/alanmcgovern/Mono.Nat/issues/17