[core] Mark device lost as obsolete
Also hide it from intellisense so no new code uses it.
A 'lost' device is a pretty funky concept when there is
no permanent connection with the nat device.

It's pretty much impossible to reliably detect if/when
the device goes offline, or if the port mappings are
lost/reset/delete.

As such, don't pretend to offer a way to detect when
the nat device goes offline as it'll be unreliable
at best.