[core] Remove the possibility of an infinite loop
If the user passes in a message for Mono.Nat to parse,
do not emit an 'UnknownDeviceFound' event if we cannot
parse a valid device from the message.

We should be assuming the user knows what they're doing,
and if they made a mistake and passed in junk we should
ignore it.