[core] Remove an unsafe way of cancelling operations
If the socket operations do not natively support cancellation,
our only option is to Close/Dispose the socket if we wish to
add cancellation.

This approach simply causes the library to stop waiting for the
socket operation to complete, which (in turn) adds a race condition
where the underlying socket could be accessed in a thread unsafe
way.