[core] Remove some async/await to simplify things
As the library was recently refactored to allow the listeners
to be re-instantiated when someone calls NatUtility.StopDiscovery
and then NatUtility.StartDiscovery, it is now easier to handle
cleanup.

Rather than relying on await async tasks in potentially blocking
ways, we can simply dispose the underlying sockets. This will
interrupt any pending send/receive calls safely.

As a result, we can remove some of the complexity surrounding
cancellation.